Refactor AesExtensions
