+++ src/SampleTest/Security.Cryptography/AesTest.cs
@@ -1,37 +1,9 @@
+using SampleLib.Security;
 using System.Security.Cryptography;
 
 namespace SampleTest.Security.Cryptography;
 
 public class AesTest {
-	// 暗号化
-	private static byte[] Encrypt(Aes aes, string plain) {
-		using var memoryStream = new MemoryStream();
-		using var cryptoStream = new CryptoStream(
-			stream: memoryStream,
-			transform: aes.CreateEncryptor(),
-			mode: CryptoStreamMode.Write,
-			leaveOpen: false);
-
-		using var writer = new StreamWriter(cryptoStream);
-		writer.Write(plain);
-		writer.Close();
-
-		return memoryStream.ToArray();
-	}
-
-	// 復号
-	private static string Decrypt(Aes aes, byte[] cipher) {
-		using var memoryStream = new MemoryStream(cipher);
-		using var cryptoStream = new CryptoStream(
-			stream: memoryStream,
-			transform: aes.CreateDecryptor(),
-			mode: CryptoStreamMode.Read,
-			leaveOpen: false);
-
-		using var reader = new StreamReader(cryptoStream);
-		return reader.ReadToEnd();
-	}
-
 	private static string GetHexString(byte[] bytes) {
 		return BitConverter.ToString(bytes).ToLower().Replace("-", "");
 	}
@@ -49,11 +21,11 @@ public class AesTest {
 
 		// Act
 		// 暗号化
-		var cipher = Encrypt(aes, "あいうえお");
+		var cipher = aes.Encrypt("あいうえお");
 		_output.WriteLine(GetHexString(cipher));
 
 		// 復号
-		var actual = Decrypt(aes, cipher);
+		var actual = aes.Decrypt(cipher);
 
 		// Assert
 		Assert.Equal("あいうえお", actual);
@@ -69,9 +41,9 @@ public class AesTest {
 		aes2.IV = aes1.IV;
 
 		// Act
-		var cipher = Encrypt(aes1, "あいうえお");
+		var cipher = aes1.Encrypt("あいうえお");
 		_output.WriteLine(GetHexString(cipher));
-		var actual = Decrypt(aes2, cipher);
+		var actual = aes1.Decrypt(cipher);
 
 		// Assert
 		Assert.Equal("あいうえお", actual);
